Ensure that background callback expiry is communicated in MCP tool
